refactor: :recycle: refactor feedback form handling logic for improved validation and submission
